Computers when moving now render the "move" (slide) animation.

+++ src/main/java/org/terasology/computer/module/inventory/RelativeInventoryBindingCustomObject.java
@@ -23,6 +23,7 @@ import org.terasology.logic.inventory.InventoryAccessComponent;
 import org.terasology.logic.inventory.InventoryComponent;
 import org.terasology.math.Direction;
 import org.terasology.math.IntegerRange;
+import org.terasology.math.geom.Vector3f;
 import org.terasology.math.geom.Vector3i;
 import org.terasology.world.BlockEntityRegistry;
 
@@ -56,7 +57,7 @@ public class RelativeInventoryBindingCustomObject implements CustomObject, Inven
 
     @Override
     public InventoryWithSlots getInventoryEntity(int line, ComputerCallback computerCallback) throws ExecutionException {
-        Vector3i computerLocation = computerCallback.getComputerLocation();
+        Vector3f computerLocation = computerCallback.getComputerLocation();
         Vector3i directionVector = inventoryDirection.getVector3i();
         Vector3i inventoryLocation = new Vector3i(
                 computerLocation.x + directionVector.x,
